requirements.txt
altair==4.2.2
folium==0.14.0
numpy==1.21.5
pandas==1.4.4
Pillow==9.5.0
plotly==5.9.0
pyttsx3==2.90
requests==2.28.1
streamlit_folium==0.11.1
streamlit_lottie==0.0.3
